nervige warnings entfernt

+++ Uebung5/skeleton-cg1_ex4 - Friedrich2/Texture.hpp
@@ -63,6 +63,6 @@ namespace World{
   void setMaterial();
 
 
-  void raytrace(int x, int y);
+  void raytrace(float x, float y);
   Ray getRay(int x, int y, float xPixel, float yPixel);
 };
